BONUS WHERE TO USE PLOTS
How this book is organized
This book is divided into chapters, one for each of the main categories for 
using a data visualization. Each chapter is headed by a short introduction, 
followed by a list of chart types falling in that category. Each chart type is 
accompanied by a short description and one or more icons.